streamlit_app/requirements.txt
streamlit>=1.35
pymongo>=4.7
python-dotenv>=1.0
pydantic>=2.6
ulid-py>=1.1
pandas>=2.2
numpy>=1.26
plotly>=5.22
pyarrow>=16.0
pytz>=2024.1
bcrypt>=4.1
